backend/requirements.txt

# FastAPI framework for building APIs
fastapi==0.111.0

# Uvicorn ASGI server with standard extras for running FastAPI
uvicorn[standard]==0.23.2

# SQLAlchemy ORM for database interactions
sqlalchemy==2.0.26

# PostgreSQL database adapter for Python
psycopg2-binary==2.9.6

# Python JOSE library for JWT token creation and verification
python-jose==3.3.0

# Passlib library for password hashing
passlib[bcrypt]==1.7.4

# Required for handling file uploads (version constraint removed to resolve conflict)
python-multipart

# Pydantic for data validation and settings management
pydantic==2.6.2